+++ codegen/cerata/src/cerata/node_array.h
@@ -21,6 +21,7 @@
 #include <deque>
 
 #include "cerata/node.h"
+#include "cerata/port.h"
 #include "cerata/edge.h"
 #include "cerata/type.h"
 
@@ -92,9 +93,10 @@ class PortArray : public NodeArray, public Term {
  public:
   /// @brief Get a smart pointer to a new ArrayPort.
   static std::shared_ptr<PortArray> Make(const std::string &name,
-                                         std::shared_ptr<Type> type,
+                                         const std::shared_ptr<Type>& type,
                                          std::shared_ptr<Node> size,
-                                         Port::Dir dir = Port::Dir::IN);
+                                         Port::Dir dir = Port::Dir::IN,
+                                         const std::shared_ptr<ClockDomain>& domain = default_domain());
 
   /// @brief Get a smart pointer to a new ArrayPort with a base type other than the default Port.
   static std::shared_ptr<PortArray> Make(const std::shared_ptr<Port> &base_node,
@@ -105,7 +107,9 @@ class PortArray : public NodeArray, public Term {
 
  protected:
   /// @brief Construct a new port array.
-  PortArray(const std::shared_ptr<Port> &base, std::shared_ptr<Node> size, Term::Dir dir);
+  PortArray(const std::shared_ptr<Port> &base,
+            std::shared_ptr<Node> size,
+            Term::Dir dir);
 };
 
 }  // namespace cerata
